Move to Axiom Logger
