+++ public/js/plugin.js
@@ -1,4 +1,4 @@
-import log from "./logger";
+import log from "../../utils/logger.js";
 
 ("use strict");
 class CPlugin {
