+++ public/js/viewer.js
@@ -1,7 +1,7 @@
 import { h, render } from "https://esm.sh/preact";
 import { useEffect, useRef, useState } from "https://esm.sh/preact/hooks";
 import htm from "https://esm.sh/htm";
-import log from "./logger.js";
+import log from "../../utils/logger.js";
 const html = htm.bind(h);
 
 function App() {
